chartsdb/requirements.txt
2dj-database-url==0.4.2
Django==1.11.4
django-crispy-forms==1.6.1
djangorestframework==3.6.3
gunicorn==19.7.1
mysqlclient==1.3.10
olefile==0.44
Pillow==4.2.1
psycopg2==2.7.3
Pygments==2.2.0
pytz==2017.2
